separate main.js on modules
